+++ code/system.ui/react.common/src/ui/PropList/u.format.ts
