+++ code/system.ui/react.common/src/ui/PropList/ui.Item/common.ts
@@ -1,2 +1,2 @@
 export * from '../common';
-export { Wrangle } from '../util';
+export { Wrangle } from '../u';
